requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
itsdangerous
bcrypt
requests
beautifulsoup4
psycopg2-binary
fpdf2
webauthn>=2.0.0
pywebpush>=1.14.0
cryptography>=42.0.0
http-ece>=1.1.0
boto3>=1.34.0
alembic>=1.13.0
redis>=5.0.0
python-dateutil>=2.9.0
python-dotenv
